# modules/mini-chat/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mini Chat API",
    "version": "1.0.0",
    "description": "Multi-tenant AI chat module. Provides conversational AI with streaming responses, document and image upload, file search, per-user quota enforcement, and audit event emission. All endpoints require authentication and tenant license verification (ai_chat feature)."
  },
...
  "servers": [],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/chats/{id}/turns/{request_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ChatId"
        },
        {
          "$ref": "#/components/parameters/RequestId"
        }
      ],
      "get": {
        "operationId": "getTurnStatus",
        "tags": ["turns"],
        "summary": "Get authoritative turn status",
        "description": "Returns the authoritative server-side turn state for a given request_id. Used by the UI for disconnect recovery. Internal states map to API states: running->running, completed->done, failed->error, cancelled->cancelled. Includes error_code (when state is 'error') and assistant_message_id (when state is 'done') so clients can resolve lifecycle uncertainty after disconnect without listing or scanning message history. Retrieving the assistant message content itself may require one follow-up request. Billing outcome and provider internals are not exposed.",
        "responses": {
          "200": {
            "description": "Turn status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnStatus"
                },
                "example": {
                  "request_id": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
                  "state": "done",
                  "error_code": null,
                  "assistant_message_id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                  "updated_at": "2025-06-15T10:36:30Z"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "editTurn",
        "tags": ["turns"],
        "summary": "Edit the last user turn and regenerate response",
        "description": "Replaces the content of the last user message and generates a new assistant response via SSE. The previous turn is soft-deleted with `replaced_by_request_id` set. Original `attachment_ids` are preserved. Only the most recent non-deleted turn in a terminal state may be edited.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTurnRequest"
              },
              "example": {
                "content": "What about Q3 profit margins?"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream with new assistant response. Same event contract as sendMessage.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events stream. See sendMessage operation for event types."
                }
              }
            }
          },
          "400": {
            "description": "Turn not in terminal state. Code: `invalid_turn_state`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Target is not the most recent turn. Code: `not_latest_turn`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded for regeneration. Code: `quota_exceeded` (`quota_scope`: `tokens`) for user quota exhaustion (always pre-stream), or `rate_limited` for upstream provider throttling. `rate_limited` may also appear as a terminal SSE `event: error` mid-stream (HTTP 200). Same semantics as sendMessage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Image bytes exceeded per-turn limit. Code: `image_bytes_exceeded`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Effective model does not support multimodal input. Code: `unsupported_media`. Same semantics as sendMessage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "LLM provider error. Code: `provider_error`. Same semantics as sendMessage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "LLM provider timeout. Code: `provider_timeout`. Same semantics as sendMessage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTurn",
        "tags": ["turns"],
        "summary": "Delete the last turn",
        "description": "Soft-deletes the most recent turn (user message + assistant response). Only the latest non-deleted turn in a terminal state may be deleted. No replacement turn is created.",
        "responses": {
          "200": {
            "description": "Turn deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnDeleteResponse"
                },
                "example": {
                  "request_id": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
                  "deleted": true
                }
              }
            }
          },
          "400": {
            "description": "Turn not in terminal state. Code: `invalid_turn_state`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Target is not the most recent turn. Code: `not_latest_turn`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chats/{id}/turns/{request_id}:retry": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ChatId"
        },
        {
          "$ref": "#/components/parameters/RequestId"
        }
      ],
      "post": {
        "operationId": "retryTurn",
        "tags": ["turns"],
        "summary": "Retry the last turn",
        "description": "Soft-deletes the previous turn and re-submits the original user message (including any attachment_ids) for a new assistant response via SSE. Only the most recent non-deleted turn in a terminal state may be retried.",
        "responses": {
          "200": {
            "description": "SSE stream with new assistant response. Same event contract as sendMessage.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events stream. See sendMessage operation for event types."
                }
              }
            }
          },
          "400": {
            "description": "Turn not in terminal state. Code: `invalid_turn_state`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Target is not the most recent turn. Code: `not_latest_turn`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded for regeneration. Code: `quota_exceeded` (`quota_scope`: `tokens`) for user quota exhaustion (always pre-stream), or `rate_limited` for upstream provider throttling. `rate_limited` may also appear as a terminal SSE `event: error` mid-stream (HTTP 200). Same semantics as sendMessage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Platform AuthN JWT or opaque token. Extracts user_id and tenant_id into the request security context."
      }
    },
    "parameters": {
      "ChatId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Chat UUID.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "RequestId": {
        "name": "request_id",
        "in": "path",
        "required": true,
        "description": "Client-generated idempotency key UUID identifying a turn.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "Forbidden": {
        "description": "Tenant lacks `ai_chat` license (code: `feature_not_licensed`) or subject lacks permission (code: `insufficient_permissions`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found or not accessible under current authorization constraints. Code: `chat_not_found` or `message_not_found`. AuthZ fail-closed: denied requests with a concrete resource ID return 404.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "description": "Standard error envelope for JSON error responses and SSE terminal error events. Error messages and codes MUST NOT contain provider-scoped identifiers (provider_file_id, vector_store_id, deployment names). All error codes are stable internal codes.",
        "oneOf": [
          { "$ref": "#/components/schemas/QuotaExceededError" },
          { "$ref": "#/components/schemas/GenericError" }
        ]
      },
      "GenericError": {
        "type": "object",
        "required": ["code", "message"],
        "description": "Error response for all non-quota error codes. No provider identifiers in any field.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Canonical error code.",
            "enum": [
              "invalid_request",
              "feature_not_licensed",
              "insufficient_permissions",
              "chat_not_found",
              "message_not_found",
              "model_not_found",
              "generation_in_progress",
              "request_id_conflict",
              "rate_limited",
              "web_search_disabled",
              "file_too_large",
              "unsupported_file_type",
              "too_many_images",
              "image_bytes_exceeded",
              "unsupported_media",
              "provider_error",
              "provider_timeout",
              "not_latest_turn",
              "invalid_turn_state",
              "invalid_reaction_target"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error description. MUST NOT contain provider-scoped identifiers (provider_file_id, vector_store_id, deployment names, upstream error bodies)."
          }
        }
      },
      "QuotaExceededError": {
        "type": "object",
        "required": ["code", "message", "quota_scope"],
        "description": "Error response for quota exhaustion. The quota_scope field is REQUIRED and distinguishes token quota exhaustion, upload quota exhaustion, and web search quota exhaustion. No provider identifiers in any field.",
        "properties": {
          "code": {
            "type": "string",
            "enum": ["quota_exceeded"]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error description. MUST NOT contain provider-scoped identifiers."
          },
          "quota_scope": {
            "type": "string",
            "enum": ["tokens", "uploads", "web_search"],
            "description": "Distinguishes token quota exhaustion (tokens), upload quota exhaustion (uploads), and web search quota exhaustion (web_search). Clients MUST use this field, not message parsing, to determine quota scope."
          }
        }
      },
...
      "TurnStatus": {
        "type": "object",
        "required": ["request_id", "state", "updated_at"],
        "description": "Authoritative server-side turn state for lifecycle resolution and disconnect recovery. Internal-to-API state mapping: running->running, completed->done, failed->error, cancelled->cancelled. Terminal states (done, error, cancelled) are immutable — once reached, the state never changes. Exactly one terminal outcome exists per (chat_id, request_id) pair. All terminal states indicate that quota settlement has been finalized atomically with the state transition. A state of 'done' guarantees that the full assistant message content is durably persisted and available for idempotent replay. Optional fields (error_code, assistant_message_id) provide actionable context so clients can resolve lifecycle uncertainty after disconnect without listing or scanning message history; retrieving the assistant message content may require one follow-up request. Billing outcome and provider internals are not exposed. chat_id is not included in the response body because it is already present in the URL path.",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": ["running", "done", "error", "cancelled"],
            "description": "Turn state. 'done', 'error', 'cancelled' are terminal and immutable. 'error' indicates the turn failed — check error_code to determine if it was a provider failure (provider_error, provider_timeout) or a system timeout (orphan_timeout). 'cancelled' indicates client disconnect or internal abort — no SSE error event was emitted. Use this endpoint to resolve state after disconnect. IMPORTANT: error_code='orphan_timeout' means stream ended without provider terminal event (system timeout, NOT provider error); display 'Request timed out' to users, NOT 'Provider error'."
          },
          "error_code": {
            "type": "string",
            "nullable": true,
            "description": "Terminal error code when state is 'error'. Disambiguates failure cause. Common values: 'provider_error' (LLM provider returned terminal error), 'provider_timeout' (provider request timed out), 'orphan_timeout' (turn stuck beyond watchdog timeout — system timeout, NOT provider failure; stream ended without provider terminal event), 'context_length_exceeded' (preflight rejection), 'quota_overshoot_exceeded' (actual usage exceeded overshoot tolerance). Null for non-error terminal states and while running. CRITICAL: 'orphan_timeout' indicates a SYSTEM timeout (pod crash, network partition, orphan watchdog cleanup), not a provider-side error. UX SHOULD display 'Request timed out. Please try again.' for orphan_timeout, NOT 'Provider error.' Support and analytics: orphan_timeout metrics should be tracked separately from provider_error to distinguish infrastructure issues from LLM provider issues. Mapped from internal chat_turns.error_code. Provider identifiers and billing outcome are not exposed."
          },
          "assistant_message_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Persisted assistant message ID when state is 'done'. Null while running, on error, or on cancellation. Clients can use this to fetch the assistant message via `GET /v1/chats/{id}/messages?$filter=id eq '{assistant_message_id}'` without scanning full history."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TurnDeleteResponse": {
        "type": "object",
        "required": ["request_id", "deleted"],
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
      "EditTurnRequest": {
        "type": "object",
        "required": ["content"],
        "description": "Request body for editing the last user turn. Original attachment_ids are preserved automatically.",
        "properties": {
          "content": {
            "type": "string",
            "description": "Replacement user message text."
          }
        }
      },
...
    }
  }
}